NB1: Switch to prebuilt health HAL
Unfortunately, it seems Nokia went with some proprietary method to implement
Quick charging that has modifications to the charger blob that also depends
on proprietary changes to the health HAL, so to prevent QC from causing
issues on our device, especially when offline charging? Let's switch to
prebuilt health HAL and allow it to work as Nokia inteded it.